refactor: #151 LogController 인터페이스화
